move resource btn to component
